Add utils.go
